requirements.txt
beautifulsoup4
requests
streamlit>=1.20.0
altair<5
psycopg2-binary
